Create summary reports of receiver project data from the OTN data push

Create summary reports of receiver project data from the OTN data push

Usage

otn_tag_summary(project = NULL, server = NULL, matched = NULL, ...)

Arguments

project

Character. The project code.

server

Character. URL of the OTN-style Plone server. Defaults to the main OTN server at https://members.oceantrack.org.

matched

Default is NULL: OTN matched detections in parquet format will be downloaded from the data server. If you do not wish to download your files or your node does not use an OTN-style Plone CMS, this argument also accepts a character vector of file paths of your matched detections. These can be parquet, CSVs, or zipped folders.

...

Arguments passed to otndo::make_tag_push_summary

matos_project

MATOS project number or name that you wish to have summarized

Examples

if (FALSE) { # \dontrun{
# You can just use your ACT project number
otn_tag_summary("mdwea")

# Or provide an optional date to summarize "What's New".
otn_tag_summary("mdwea", since = "2025-11-01")
} # }